requirement.txt
streamlit
streamlit-option-menu
transformers
diffusers
soundfile
pillow
numpy
